split up json persistence into JsonUtil.scala
